OC-11428: code change based on code review
