+++ web/src/main/java/org/akaza/openclinica/service/ParticipantService.java
@@ -20,10 +20,7 @@ public interface ParticipantService {
 
 	 List<StudySubjectBean> getStudySubject(StudyBean study);
 	 StudyBean validateRequestAndReturnStudy(String studyOid, String siteOid,HttpServletRequest request);
-	 UserAccountBean getUserAccount(HttpServletRequest request);
-	void processBulkParticipants(StudyBean study, String studyOid, StudyBean siteStudy, String siteOid,
-								 UserAccountBean user, String accessToken, String customerUuid, MultipartFile file,
-								 JobDetail jobDetail, Locale locale, String uri, Map<String, Object> map) throws Exception;
+	 UserAccountBean getUserAccount(HttpServletRequest request);	
 	
 	StudyParticipantDetailDTO buildStudyParticipantDetailDTO(StudySubject studySubject);
 }
