+++ web/src/main/java/org/akaza/openclinica/service/StudyParticipantServiceImpl.java
@@ -16,15 +16,11 @@ import org.akaza.openclinica.dao.hibernate.StudySubjectDao;
 import org.akaza.openclinica.dao.hibernate.SubjectDao;
 import org.akaza.openclinica.dao.managestudy.StudyDAO;
 import org.akaza.openclinica.dao.managestudy.StudySubjectDAO;
+import org.akaza.openclinica.dao.service.StudyParameterValueDAO;
 import org.akaza.openclinica.dao.submit.SubjectDAO;
 import org.akaza.openclinica.domain.datamap.*;
 import org.akaza.openclinica.domain.enumsupport.JobType;
-import org.akaza.openclinica.domain.user.UserAccount;
-import org.akaza.openclinica.exception.OpenClinicaException;
 import org.akaza.openclinica.exception.OpenClinicaSystemException;
-import org.akaza.openclinica.i18n.util.ResourceBundleProvider;
-import org.akaza.openclinica.service.crfdata.ErrorObj;
-import org.akaza.openclinica.validator.ParticipantValidator;
 import org.akaza.openclinica.web.restful.errors.ErrorConstants;
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.slf4j.Logger;
@@ -36,16 +32,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.PrintWriter;
-import java.io.UnsupportedEncodingException;
 import java.text.SimpleDateFormat;
 import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
-import static org.akaza.openclinica.service.UserService.BULK_JOBS;
-import static org.akaza.openclinica.service.UserServiceImpl.SEPERATOR;
 
 /**
  * This Service class is used with Add Participant Rest Api
@@ -86,8 +76,7 @@ public class StudyParticipantServiceImpl implements StudyParticipantService {
 
     private StudyDAO studyDao;
     private StudySubjectDAO studySubjectDao;
-    private SubjectDAO subjectDao;
-    private ParticipantValidator participantValidator;
+    private SubjectDAO subjectDao;  
 
     SimpleDateFormat sdf_fileName = new SimpleDateFormat("yyyy-MM-dd'-'HHmmssSSS'Z'");
     public static final String DASH = "-";
@@ -99,7 +88,8 @@ public class StudyParticipantServiceImpl implements StudyParticipantService {
     public static final String EMAIL_PATTERN = "^[A-Za-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[A-Za-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[A-Za-z0-9](?:[A-Za-z0-9-]*[A-Za-z0-9])?\\.)+[A-Za-z0-9](?:[A-Za-z0-9-]*[A-Za-z0-9])?$";
 
 
-
+    private  StudyBean currentStudy;
+    private  StudyBean siteStudy;
 
     public AddParticipantResponseDTO addParticipant(AddParticipantRequestDTO addParticipantRequestDTO, UserAccountBean userAccountBean, String studyOid, String siteOid , String customerUuid, ResourceBundle textsBundle,String accessToken, String register ) {
         boolean createNewParticipant=false;
@@ -107,11 +97,10 @@ public class StudyParticipantServiceImpl implements StudyParticipantService {
         Study tenantSite = studyHibDao.findByOcOID(siteOid);
         StudyBean tenantStudyBean = getStudyDao().findByOid(studyOid);
         StudyBean tenantSiteBean = getStudyDao().findByOid(siteOid);
-
-        participantValidator = this.getParticipantValidator();
-        participantValidator.setCurrentStudy(tenantStudyBean);
-        participantValidator.setSiteStudy(tenantSiteBean);
-        if (participantValidator.isEnrollmentCapped())
+       
+        currentStudy = tenantStudyBean;
+        siteStudy = tenantSiteBean;
+        if (isEnrollmentCapped())
             throw new OpenClinicaSystemException( ErrorConstants.ERR_PARTICIPANTS_ENROLLMENT_CAP_REACHED);
         
         if (StringUtils.isEmpty(addParticipantRequestDTO.getSubjectKey()))
@@ -334,20 +323,55 @@ public class StudyParticipantServiceImpl implements StudyParticipantService {
     }
 
 
-	public ParticipantValidator getParticipantValidator() {
-		if(participantValidator != null) {
-			return participantValidator;
-		}else {
-			participantValidator = new ParticipantValidator(dataSource);
-		}
-		
-		return participantValidator;
-		
-	}
+    public boolean isEnrollmentCapped(){
 
+        boolean capIsOn = isEnrollmentCapEnforced();
 
-	public void setParticipantValidator(ParticipantValidator participantValidator) {
-		this.participantValidator = participantValidator;
-	}
+        StudySubjectDAO studySubjectDAO = this.getStudySubjectDao();
+        int numberOfSubjects = studySubjectDAO.getCountofActiveStudySubjects();
+
+        StudyDAO studyDAO = this.getStudyDao();
+        StudyBean sb = null;
+        if(currentStudy.getParentStudyId()!=0){
+            sb = (StudyBean) studyDAO.findByPK(currentStudy.getParentStudyId());
+        }else{
+             sb = (StudyBean) studyDAO.findByPK(currentStudy.getId());
+        }
+        int  expectedTotalEnrollment = sb.getExpectedTotalEnrollment();
+
+        if (numberOfSubjects >= expectedTotalEnrollment && capIsOn)
+            return true;
+        else
+            return false;
+    }
+
+    /**
+     * if it's site level, then also need to check study 
+     * @return
+     */
+    private boolean isEnrollmentCapEnforced(){
+        StudyParameterValueDAO studyParameterValueDAO = new StudyParameterValueDAO(this.dataSource);
+       
+        boolean capEnforcedSite = false;
+        String  enrollmentCapStatusSite = null;
+        
+        String enrollmentCapStatus = studyParameterValueDAO.findByHandleAndStudy(currentStudy.getId(), "enforceEnrollmentCap").getValue();
+        boolean capEnforced = Boolean.valueOf(enrollmentCapStatus);
+        
+        // check at the site level
+        if(siteStudy != null) {
+        	int siteId = siteStudy.getId();
+        	enrollmentCapStatusSite = studyParameterValueDAO.findByHandleAndStudy(siteId, "enforceEnrollmentCap").getValue();
+        	capEnforcedSite = Boolean.valueOf(enrollmentCapStatusSite);        	
+        }
+        
+        if(capEnforcedSite || capEnforced) {
+        	return true;
+        }else {
+        	return false;
+        }
+        
+    }
+
 
 }
